# Speed up and enhance reading from (and writing to) a GRASS GIS database