--- a/DESCRIPTION
+++ b/DESCRIPTION
@@ -18,5 +18,6 @@ SystemRequirements: GRASS (>= 7)
 License: GPL (>= 2)
 URL: https://rsbivand.github.io/rgrass/, https://grass.osgeo.org/, https://github.com/rsbivand/rgrass, https://lists.osgeo.org/mailman/listinfo/grass-stats
 BugReports: https://github.com/rsbivand/rgrass/issues/
-Collate: AAA.R options.R rgrass.R rast_link.R vect_link.R vect_link_ng.R initGRASS.R xml1.R
+Collate: AAA.R options.R rgrass.R rast_link.R vect_link.R vect_link_ng.R 
+  initGRASS.R xml1.R gdal_grass.R