--- a/DESCRIPTION
+++ b/DESCRIPTION
@@ -19,5 +19,5 @@ License: GPL (>= 2)
 URL: https://rsbivand.github.io/rgrass/, https://grass.osgeo.org/, https://github.com/rsbivand/rgrass, https://lists.osgeo.org/mailman/listinfo/grass-stats
 BugReports: https://github.com/rsbivand/rgrass/issues/
 Collate: AAA.R options.R rgrass.R rast_link.R vect_link.R vect_link_ng.R 
-  initGRASS.R xml1.R gdal_grass.R
+  initGRASS.R xml1.R gdal_grass.R read_helpers.R