This PR attempts to address the following:

- **`read_RAST()`**, **`read_VECT()`**: implement the standalone [GDAL-GRASS drivers](https://github.com/OSGeo/gdal-grass) by default. This avoids exporting a GRASS GIS layer to a temporary file before reading into R. As a result one will get higher reading rates; this effect will increase with the size of the layer. Notes:
  - In `read_RAST()` this has only been implemented for the **terra** case, as it is the default.
  - Some issues still exist with the driver, such as a missing EPSG code. This is referred in the documentation. However the data are successfully returned.
  - The `v.out.ogr` and `r.out.gdal` pathways (writing a temporary file) remain in place when the necessary GDAL-GRASS driver is missing or if the `use_gdal_grass_driver` argument is set as `FALSE`.
  - in case the **terra** object already refers to a GRASS GIS database layer (i.e. has been created using the GDAL-GRASS driver), `write_RAST()` and `write_VECT()` will prevent calls to GRASS GIS to import such a source, as this would not succeed.
  - this closes #75.
- `read_VECT()`: provide support for the **`SpatVectorProxy`** class by adding a `proxy` argument (see [`terra::query()`](https://rspatial.github.io/terra/reference/query.html) to use `SpatVectorProxy` in R).
- arguments **`ignore.stderr`** and **`Sys_ignore.stdout`**: implement them in the default `execGRASS()` approach (uses `system2()`) – they were only implemented in the legacy approach – and make sure both arguments are exposed by `read_RAST()` and `read_VECT()`.
- **`write_VECT()`**: when the `SpatVector` object refers to a source file, a temporary GeoPackage is no longer written and the file is used directly by `v.in.ogr`. A similar shortcut was already implemented for `write_RAST()`.

Further testing, especially with Windows and macOS are most welcome!

A small benchmark is shown below.

<details><summary><strong>Code to create the benchmark</strong></summary>

``` r
library(dplyr, warn.conflicts = FALSE)
library(rgrass)
#> GRASS GIS interface loaded with GRASS version: (GRASS not running)
grass_conf <- initGRASS(
  gisBase = "/usr/lib/grass83",
  gisDbase = "/home/floris/grassdata",
  location = "nc_basic_spm_grass7",
  mapset = "PERMANENT",
  override = TRUE
)
vec_without_gdalgrass <- bench::mark(
  vec_without_gdalgrass = read_VECT(
    "streets",
    use_gdal_grass_driver = FALSE,
    ignore.stderr = TRUE
  ),
  iterations = 4
)
suppressWarnings({
  vec_with_gdalgrass <- bench::mark(
    vec_with_gdalgrass = read_VECT(
      "streets",
      use_gdal_grass_driver = TRUE
    ),
    iterations = 4
  )
})
rast_without_gdalgrass <- bench::mark(
  rast_without_gdalgrass = read_RAST(
    "elevation_shade",
    use_gdal_grass_driver = FALSE,
    ignore.stderr = TRUE
  ),
  iterations = 10
)
rast_with_gdalgrass <- bench::mark(
  rast_with_gdalgrass = read_RAST(
    "elevation_shade",
    use_gdal_grass_driver = TRUE
  ),
  iterations = 10
)

bind_rows(
  vec_without_gdalgrass,
  vec_with_gdalgrass,
  rast_without_gdalgrass,
  rast_with_gdalgrass
) |> 
  select(expression, min, median, iterations = n_itr) |> 
  knitr::kable()
```

</details>

| expression             |      min |   median | iterations |
|:-----------------------|---------:|---------:|-----------:|
| vec_without_gdalgrass  |   12.92s |   13.09s |          3 |
| vec_with_gdalgrass     |    4.13s |    4.18s |          4 |
| rast_without_gdalgrass | 476.39ms | 484.12ms |          9 |
| rast_with_gdalgrass    | 143.08ms | 146.35ms |          9 |

<sup>Created on 2024-06-16 with [reprex v2.1.0](https://reprex.tidyverse.org)</sup>

Some system specs:

```bash
$ ogrinfo --version
GDAL 3.8.4, released 2024/02/08
$ 
$ apt policy libgdal-grass
libgdal-grass:
  Installed: 1:1.0.2-7+jammy3
  Candidate: 1:1.0.2-7+jammy3
  Version table:
 *** 1:1.0.2-7+jammy3 500
        500 http://ppa.launchpad.net/ubuntugis/ubuntugis-unstable/ubuntu jammy/main amd64 Packages
        100 /var/lib/dpkg/status
     3.4.1-3 500
        500 http://ftp.belnet.be/ubuntu jammy/universe amd64 Packages
$ 
$ inxi -Sxxx
System:
  Host: xxx Kernel: 5.15.0-112-generic x86_64 bits: 64 compiler: gcc
    v: 11.4.0 Desktop: Cinnamon 6.0.4 tk: GTK 3.24.33 wm: muffin vt: 7
    dm: LightDM 1.30.0 Distro: Linux Mint 21.3 Virginia
    base: Ubuntu 22.04 jammy
```