@hellik 2024-06-16
Point 1 by @rsbivand is an important one regarding winGRASS. Different GDAL versions built by different compilers may interfere. That's a packaging issue.

Though I find it an interesting approach worth to be followed.

The fallbacks are the methods currently used?