@florisvdh 2024-06-17
> The fallbacks are the methods currently used?

@hellik yes, if GDAL is not aware of the GDAL-GRASS driver (so indeed it needs to be set up correctly), the current implementation is used. In that case GRASS GIS is called with `r.out.gdal` or `v.out.ogr` to first write a temporary file either with the RRASTER (raster) or the GPKG (vector) GDAL driver, respectively.

For a binary driver installation that works out of the box Ubuntu-based systems, one can use the ubuntugis-unstable PPA, which is where the official GRASS GIS release lives, and install `libgdal-grass`. It is on par with its GRASS GIS and GDAL dependencies in the same repo. Still, to be user-friendly and work out of the box on all systems, it would indeed be great to have it natively included in GDAL.

These are the driver issues encountered a few days ago and posted at the driver repo, with help of @neteler: https://github.com/OSGeo/gdal-grass/issues/46, https://github.com/OSGeo/gdal-grass/issues/47, https://github.com/OSGeo/gdal-grass/issues/48.