@florisvdh 2024-06-17
> suggest that it can write

No, it's only relevant to `read_RAST()` and `read_VECT()`, since indeed it's a read-only driver. So in this PR the driver functionality is only added in `read_RAST()` and `read_VECT()`.

In relation to the driver, `write_*()` functions were only updated to refuse objects that already refer to a GRASS data source.

Apart from that, this PR also adds some `read_*()` and `write_*()` functionality that is agnostic to which approach is being used (see PR description).

> It will not be available on platforms where terra is not built with driver-enabled GDAL.

GDAL needs no specific configuration to use the driver; it uses its autoload mechanism to detect the driver and make it available. See https://github.com/OSGeo/gdal-grass/blob/main/README.md:

> _With this driver package you configure and install GDAL normally, then build and install GRASS normally and finally build and install this driver in [GDAL's "autoload" directory](https://gdal.org/user/configoptions.html#driver-management)._

On my system it was sufficient for {terra} to have been compiled against the GDAL version that was in place before the additional driver was installed. Also no additional `GDAL_DRIVER_PATH` environment variable (as documented at https://github.com/OSGeo/gdal-grass/blob/main/README.md) was needed for the driver to be detected, since the binary driver from the ubuntugis-unstable PPA is installed in an expected path:

```bash
$ locate gdalplugins | grep GRASS
/usr/lib/x86_64-linux-gnu/gdalplugins/gdal_GRASS.so
/usr/lib/x86_64-linux-gnu/gdalplugins/ogr_GRASS.so
```