@florisvdh 2024-06-17
> Running `rgrass` built from the `gdal_grass` branch fails `CMD check` with this log: [00check.log](https://github.com/user-attachments/files/15864349/00check.log)

Thanks. It's introduced by 682c7fc, 'needed' in order to not get the same error on my system, where there's no `rsb` mapset.

So since this error now pops up on your side I guess on your machine the `rsb` mapset will be the active mapset, so that `write_RAST()` writes the layers there, invoking the error.

The solution will be to be explicit about the mapset where new layers will be written. I suggest that we create a custom mapset explicitly in the examples, so that written layers end up there, not in `PERMANENT`. Do you agree @rsbivand?