@rsbivand 2024-06-17
Yes, @florisvdh creating a mapset to write into is a good idea in the examples; will you implement it?