@florisvdh 2024-06-18
@rsbivand Setting a mapset in the examples of this PR has now been taken care of (in `readRAST.Rd` and `readVECT.Rd`).

I discovered that the `vect2neigh()` example in `readVECT.Rd` still somehow modifies the `PERMANENT`  mapset by writing 3 extra layers into it (based on the `census` dataset). This happens even though the custom mapset is active. Since I won't soon have time to digg into this function (already returning from the GRASS GIS meeting), I have outcommented its example so that `PERMANENT` remains untouched.