@florisvdh 2024-06-18
While fiddling with mapsets and regions yesterday, it was discovered that the GDAL-GRASS driver does not respect a mapset's region. It just reads out a raster layer as it is stored in the GRASS GIS database. This is an important difference with `r.out.gdal`, which applies the mapset's region settings (extent & resolution).

As this causes inconsistency between the `r.out.gdal` approach (i.e. writing a temporary file for use by {terra}) and the GDAL-GRASS driver approach (direct reading), we cannot consider uptake of the GDAL-GRASS raster driver already.

Consequently I have reverted the use of the GDAL-GRASS raster driver in this PR (a828acf).

A corresponding driver feature request has been posted at https://github.com/OSGeo/gdal-grass/issues/49, so at least awaiting a solution for that one before reconsideration. For the driver it makes sense to take into account the region, since this GDAL driver is tailored for GRASS GIS.

Thanks @neteler and @veroandreo for the discussion about this yesterday.

Note that for vector datasets the region is not considered in GRASS GIS, so only the raster driver approach has been reverted.