@florisvdh 2024-06-18
> @florisvdh looks OK now, would you like to merge?

OK, let's do that. I'll add updates to `NEWS.md` and the pkgdown website before merging.