@florisvdh 2024-06-18
@rsbivand you can of course rephrase `NEWS.md` if needed. Thanks for reviewing this PR.